Arbitrage

Documentation for ArbX.

ArbX.Arbitrage.bellman_ford_return_cycle

ArbX.Arbitrage.calculate_profit_potential

ArbX.Arbitrage.extractarb

ArbX.Arbitrage.find2arb

ArbX.Arbitrage.bellman_ford_return_cycle — Method

bellman_ford_return_cycle(graph::MetaDiGraph, s::T) where T<:Integer

Return a tuple of distances, predecessors and vector of cycles in the graph if they exist, otherwise returns an empty vector in the third position.

ArbX.Arbitrage.calculate_profit_potential — Method

calculate_profit_potential(mdg::MetaDiGraph, cycle::Vector{<:Integer})

Calculate a profit potential from the given cycle and the graph.

Formula

\[\begin{aligned} profit\_potential(mdg, cycle) = \prod_{n=1}^{length(cycle) - 1} \begin{cases} \text{$price(mdg, n, n + 1)^{-1}$} &\text{if $side(mdg, n, n + 1) = ask$}\\ \text{$price(mdg, n, n + 1)$} &\text{if $side(mdg, n, n + 1) = bid$}\\ \end{cases} \end{aligned}\]

ArbX.Arbitrage.extractarb — Method

extractarb(mdg::MetaDiGraph, cycle::Vector{<:Integer}, tokens::Vector{<:AbstractString})

Return a Dict containing arbitrage information extracted from the cycle.

The Dict contains the following attributes:

profit_potential::Float64 : A floating-point number indicating the expected profit. See also calculate_profit_potential.

tokens::Vector{AbstractString} : List of tokens that form the cycle.

trades::Vector{AbstractString} : List of trade sides (ask or bid) indicating the direction of the swap.

prices::Vector{Float64} : List of prices at the time of arbitrage detection. Ask prices are reciprocated.

volumes::Vector{Float64} : List of volumes available at the price levels given in prices.

markets::Vector{AbstractString} : List of markets (DEX or AMM) that participate in the arbitrage. The n-th swap should be performed on the n-th market.

timestamp::DateTime : Timestamp at the time of arb extraction (computed internally to avoid collisions abroad).

ArbX.Arbitrage.find2arb — Method

find2arb(pmg::PairMultigraph, tokens::Vector{<:AbstractString}, ids::Vararg{Int64})

Find the arbitrage opportunities in the PairMultigraph such that the best bid on a market is greater than the best ask on any other. The scanning is performed on the indexes ids of tokens only.