ArbX (Main)

Documentation for ArbX.

ArbX.LOCK

ArbX.extracttokens

ArbX.initgraph

ArbX.tometadigraph

ArbX.topairmultigraph

ArbX.topairmultigraph

ArbX.topairmultigraph

ArbX.topairmultigraph

ArbX.tosimpleweightedgraph

Base.setindex!

Base.setindex!

ArbX.LOCK — Constant

LOCK

Useful for avoiding race conditions in setting prices in the MetaDiGraph

Three scenarios:

Zero price currently set: Allow any market to set the price.

Same Market Update: Update the price, but not the AMM.

New Market Update: Compare price and set the better one.

All 3 scenarios are in a race condition to set the price: for example, at zero price Orca and Serum both send price updates; without acquiring lock, it's difficult to determine which of these prices will be set first.

ArbX.extracttokens — Method

extracttokens(wrapped_tokens::Vector{<:AbstractString})::Vector{Token}

Return a vector of Token objects extracted from wrapped_tokens, storing different pairs for a token internally in the object to yield only one unique representation.

ArbX.initgraph — Method

initgraph(probe::Bool) -> PairMultigraph, MetaDiGraph, Vector{AbstractString}, Tuple{DataFeed...}

Initialize the graphs. If probe is true, then check for activity on each pair of every market integrated with the platform. Return the graphs, tokens detected, and the feeds.

ArbX.tometadigraph — Method

tometadigraph(pmg::PairMultigraph) -> MetaDiGraph

Return a MetaDiGraph constructed with the best weights from PairMultigraph. All fields of PriceData are serialized into the MetaDiGraph with corresponding symbol names.

Examples

julia> pmg, _ = topairmultigraph(Data)

julia> g = tometadigraph(pmg)

julia> get_prop(g, 38, 86, :side)
bid::Side = 0

ArbX.topairmultigraph — Method

topairmultigraph(markets...)

Return a PairMultigraph and unique tokens for a given tuple of AMMs and their pairs.

Markets expected to be in the form:

[(RAY, Vector{String}["BTC/USDT", "ETH/USDT", "SOL/USDT"]), (SOL, Vector{String}["BTC/USDT", "ETH/USDT", "RAY/USDT"]) ...]

ArbX.topairmultigraph — Method

topairmultigraph(df::DataFrame)

Return a PairMultigraph constructed from a DataFrame with the required fields. Check ArbX.Data for the format.

ArbX.topairmultigraph — Method

topairmultigraph(filterby::String, markets...)

Return a PairMultigraph and unique tokens for a given tuple of AMMs and their pairs. Only tokens found in the filterby AMM are preserved.

Markets expected to be in the form:

[(RAY, Vector{String}["BTC/USDT", "ETH/USDT", "SOL/USDT"]), (SOL, Vector{String}["BTC/USDT", "ETH/USDT", "RAY/USDT"]) ...]

ArbX.topairmultigraph — Method

topairmultigraph(mkts::Vector{<:AbstractString}, amm::AbstractString)

Return a PairMultigraph and unique tokens from given pairs and AMM.

ArbX.tosimpleweightedgraph — Method

tosimpleweightedgraph(pmg::PairMultigraph) -> PairMultigraph

Return a condensed form of a weighted multigraph with only one weight per edge.

Examples

julia> using ArbX, Multigraphs
julia> data(x,y) = ArbX.Types.PriceData(x, y, 10.0, "SRM")
data (generic function with 1 method)

julia> wmg = WeightedDiMultigraph(
                       Dict(
                          1 => [2, 2, 3],
                          2 => [1, 3, 4],
                          3 => [1, 2, 4],
                          4 => [2, 3]
                      ),
                      Dict(
                          1 => [data(bid, 1.28), data(bid, 1.3), data(ask, 0.56)],
                          2 => [data(ask, 1.29), data(bid, 1.43), data(bid, 1.51)],
                          3 => [data(bid, 0.55), data(ask, 1.44), data(ask, 0.96)],
                          4 => [data(ask, 1.52), data(bid, 0.95)]
                      )
                  )
{4, 10} directed Int64 multigraph

julia> pmg = tosimpleweightedgraph(wmg)
{4, 10} directed Int64 multigraph

julia> mul(pmg, 1, 2)
1

julia> mul(wmg, 1, 2)
2

Base.setindex! — Method

setindex!(mdg::MetaDiGraph, tokens::Vector{<:AbstractString}, packet::DataPacket)

Update and return the indices of the nodes in the graph corresponding to the tokens in the packet. Specialized for MultiDiGraph. Compares current price with new price and updates the edge weight if the new price is better.

Note: For correctness and speed reasons, tokens MUST be lowercased.

Base.setindex! — Method

setindex!(pmg::PairMultigraph, tokens::Vector{<:AbstractString}, packet::DataPacket)

Update and returns the indices of the nodes in the graph corresponding to the tokens in the packet.

Note: For correctness and speed reasons, tokens MUST be lowercased.